Move logout, don't send email
Generally speaking should keep email as private as possible, so remove
from api

No reason for the logout endpoint to be local account specific
