+++ src/us/kbase/auth2/service/api/Logout.java
@@ -1,0 +1,69 @@
+package us.kbase.auth2.service.api;
+
+import java.util.Map;
+
+import javax.inject.Inject;
+import javax.ws.rs.CookieParam;
+import javax.ws.rs.GET;
+import javax.ws.rs.POST;
+import javax.ws.rs.Path;
+import javax.ws.rs.core.Cookie;
+import javax.ws.rs.core.NewCookie;
+import javax.ws.rs.core.Response;
+
+import org.glassfish.jersey.server.mvc.Template;
+import org.glassfish.jersey.server.mvc.Viewable;
+
+import com.google.common.collect.ImmutableMap;
+
+import us.kbase.auth2.lib.Authentication;
+import us.kbase.auth2.lib.exceptions.AuthError;
+import us.kbase.auth2.lib.exceptions.AuthenticationException;
+import us.kbase.auth2.lib.storage.exceptions.AuthStorageException;
+import us.kbase.auth2.lib.token.HashedToken;
+import us.kbase.auth2.lib.token.IncomingToken;
+
+@Path("/logout")
+public class Logout {
+
+	@Inject
+	private Authentication auth;
+	
+	@GET
+	@Template(name = "/logout")
+	public Map<String, String> logout(
+			@CookieParam("token") final String token)
+			throws AuthenticationException, AuthStorageException {
+		checkToken(token);
+		final HashedToken ht = auth.getToken(new IncomingToken(token));
+		return ImmutableMap.of("user", ht.getUserName().getName(),
+				"logouturl", "/logout/result");
+	}
+	
+	@POST
+	@Path("/result")
+	public Response logoutResult(
+			@CookieParam("token") final String token)
+			throws AuthenticationException, AuthStorageException {
+		checkToken(token);
+		final HashedToken ht = auth.revokeToken(new IncomingToken(token));
+		return Response.ok(
+				new Viewable("/logoutresult",
+						ImmutableMap.of("user", ht == null ? null :
+							ht.getUserName().getName())))
+				.cookie(new NewCookie(
+						new Cookie("token", "logout", "/", null),
+						"authtoken", 0, false))
+				.build();
+	}
+	
+	//TODO NOW make this a convenience method - API helper class
+	private void checkToken(final String token)
+			throws AuthenticationException {
+		if (token == null || token.isEmpty()) {
+			throw new AuthenticationException(AuthError.NO_TOKEN, 
+					"An authentication token must be supplied in the request.");
+		}
+	}
+	
+}
