+++ templates/logout.mustache
